refactor: move geteducationitems and makeeducationitem
